Simplify need_war_player().
Unify the handling of tile targeted actions with the handling of other
actions. The criteria for a perceived need for war is now the same for all
actions. The criteria is that the action is illegal under the current
diplomatic state but would be legal during war. No checks if another
diplomatic state than war and the current state would allow the action are
done. That could be done later in a check for non war diplomatic states.

Bombard and Attack are both unit stack targeted. Per unit war requirements
are therefore handled by the action enabler. Drop custom check.

See hrm Feature #878193