illegal_action(): fix explanation finding order.
Punishing a unit changes its state. This state change can make the action
illegal for other reasons than the real reason why it was illegal. The
reason why the action was illegal must therefore be found before the
punishment is applied.

Looking for an explanation why the action was illegal after the unit was
punished can also crash the server if the unit is killed.

This patch creates a new, smaller, bug: A client asking to not be disturbed
will end up not getting the reason why the unit was punished. Accept this
for now. (See hrm Bug #879881)

See hrm Bug #879880