Start porting the server to the Qt event loop
The server compiles again, with some refactoring to have more QString and less
char * in the client API.

The server itself was refactored to a QObject-derived class. Currently it
handles only standard input, with a few improvements with respect to the
preexisting C code:

* It now ignores EOF. You can hit Ctrl+D and nothing will happen.
* It detects non-interactive contexts and doesn't print the prompt in this
  case. Actually readline is not used at all in this case.

This includes some code that reportedly may work on Windows but it was not
tested. What will happen on Mac is another open question.