Don't try to start if the socket cannot be bound
The code was calling QCoreApplication::exit, but this is run before the event
loop is started so the call was ignored. Instead, propagate the error all the
way up to the main function and fail early.

This also means that freeciv::server can be in an invalid state. Since a socket
isn't always required (eg autogames, or if we add QLocalSocket one day), this
should be handled more gracefully.