Always call update_game_state on phase_begin
When AIs are set not to move "concurrently", each has its own phase that we
need to iterate though. The call to check_for_full_turn_done() a few lines
above makes sure that force_end_of_sniff is set when all the players in the
current phase are done moving, which is the case for AIs after begin_phase().
Then update_game_state() will move to the next phase.

Perform the call through the event loop to prevent recursion and be a little
bit less irresponsive during TC.