Reset is_new_game only after it is used
game.info.is_new_game is used to initialize a bunch of variables at the
beginning of the first turn to avoid changing the game state after loading from
a save. It is set to true for new games, and reset to false immediately
after the game starts.

It was reset too early, before it could be used, so units didn't have full MP
on game start.