Update COPYING to GPLv3
The new network code is under GPLv3+, so the aggregate cannot be used
under the GPLv2 anymore. Make it clear by updating the text in COPYING
to the GPLv3.

Reasoning:

* The existing code is under GPLv2 (or any later version). Therefore we
  can use it under GPLv3 as well (or any later version).
* Not distributing our changes under GPLv2 may terminate our rights
  under that version of the license, but we may still use the code under
  any later version.

This change does not prevents people who contribute here from also
contributing to the original GPLv2 repository. They are free to release
their modifications under GPLv2 as well as GPLv3+.