Update the city screen layout for large cities
A large number of citizens in the title box at the top could push the two side
widgets farther away from the center. Decouple the layouts.

Do the same for the units box at the bottom, so it can expand when there are
many units. Currently it doesn't set the appropriate size hint.