Reflow settings.cpp
clang-format could not understand that the bloody GEN_XXX macros ought to be at
the same level, and was indenting them to the point of rendering everything
unreadable. Move the commas auto of the macros to help clang-format, and reflow
the whole thing so it looks nice.

There was a lot of regex-based editing so this also touches other parts of the
file.