Speed up tileset loading on Windows (3x)
When loading a tileset, the client tries every possible format for every
graphics file. This is very inefficient on Windows. Make sure that png
is always tried out first, to match what all shipped files use. A rough
estimate of the speedup is 3x for hexamplio (from ~15s to ~5s in my VM).