Bump the required C++ version to C++ 17
This change is done primarily because the Qt5 shipped in Homebrew appears to be
configured to use C++ 17 and breaks the build on Travis.

C++ 17 is available on Debian Buster (already our minimal requirement for Qt),
MSYS and OS X. Using it brings nice new features without significant pitfalls.
MSVC should also support it.