Replace the random number generator
Use an implementation from the standard library. The random generator is
constrained to use a well defined algorithm and produce the same results across
implementations. Seed it by pulling the state from the high-entropy
std::random_device: it has so many bits that it's likely very hard to figure
out the state. (Of course only if the implementation of std::random_device is
correct. I trust it more than anything I could write myself.)

The main drawback is that the state might be saved in an implementation defined
way, the reference isn't very clear about this. Work around this by generating
a new, truly random seed if loading the state fails. The main reason for saving
the state is to have deterministic results when loading a save several times,
so one can't "cheat" by reloading unit a warrior kills a tank. If someone wants
to cheat, s.he will manage: firing up the editor or editing the save directly
is much easier than building a server with another standard library.

Closes #358.