Make the sentry range ruleset tweakable
The maximum distance at which a sentried unit will be woken up was hard coded
to 3 20 years ago. This is not satisfying for ruleset such as LTT with large
vision radii. A complete solution would be to wake up any sentried unit that
can see the moving unit, but it would require deeper refactoring.

As a middle ground, allow rulesets to change the sentry range from its default
of 3.

See #390.