Prevent unwanted // when downloading modpacks
When the base_url field already had a trailing / or was a relative URL (. or
..), a duplicate / was appended at the end. Some Web servers don't like URL
paths with // in them (nginx and IIS collapse them by default, Apache
apparently not).

Only append a / when the base_url would be interpreted as a file when resolving
relative URLs; base_url is always treated as a directory. (The difference is as
follows: the URL "x.txt" relative to "http://a.com/b" is "http://a.com/x.txt",
while relative to "http://a.com/b/" it is "http://a.com/b/x.txt".)

See #446.