Client positional argument to connect to URL
This allows to connect immediately using a command of the form:

    freeciv21-client user:password@host:port -a

URL syntax is easier to remember than the combination of the -n -s -p options
that was previously needed (and is still supported), and as a side-effect
allows to give a password as well. Supporting URLs also brings us closer to
setting the system handler for fc21:// links to the client (so websites can
give a fc21:// link to help users connect to a server).

Closes #468.

Internally, a single static QUrl is used to pass connection information instead
of several static field. This will allow using the path part for a
WebSocket-backed protocol and eventually WS-based encryption.