Fix a crash when auto-upgrading unique units
When Leonardo's Workshop upgrades units, it starts by collecting a list
of upgradeable units and then pick upgrades at random. This doesn't work
if upgrading one unit prevents another from being upgraded, which can
happen with unique units. Consider for instance the following upgrade
chain of unique units:

    A -> B -> C

If a player has one A and one B, and discovers the technology for C,
both the A and the B units will be candidates for Leo upgrade. Once one
of them is upgraded, the second can't be upgraded anymore because it
would create two C units. This patch fixes the handling of this corner
case in the server.