Don't quitidle if the connection is not yet established
Not-yet-established connections were considered inactive for the
quitidle server option, which is used for instance by the client.
This caused the server to exit before the client had a chance to
connect if update_game_state was called often enough, which was the case
on Windows for servers started from the client.

The fix is to consider all connections when starting the quitidle timer
instead of just the established ones.

Add a check that no connection is being attempted when the quitidle
timer fires to make sure that this doesn't happen.

See #341.