Don't block when getting an incomplete line on stdin
The server used to block when an imcomplete line was posted to stdin in
non-interactive mode. This happened on Windows when started from the
client (because of a separate issue: QWinEventNotifier goes wild when
watching the console but stdin is a pipe). Prevent the bocking by
checking whether a line can be read before attempting to read one.

With this change, the client can spawn a server on Windows and connect
to it, but the server has 100% CPU usage.

See #341.