Watch a file instead of the console on Windows
Getting events directly from the console to react to stdin breaks badly
when stdin is not a console, eg when the client starts the server. The
end result is an endless spawn of new events, making the server use 100%
CPU.

Use CreateFileA instead of GetStdHandle. CreateFileA should work under
all circumstances, not just when the input is a console.

Closes #341.