Simplify fill_basic_extra_sprite_array
Instead of duplicating the drawing logic from fill_sprite_array, call
fill_sprite_array with a virtual tile that has only the requested extra set.
Adjustments were needed to make fill_sprite_array work with this kind of tiles.

As a consequence of this change, fill_basic_extra_sprite_array now returns
several sprites for shipped tilesets. The help dialog was adjusted to take them
all into account, with the limitation that the first sprite dictates the size
of the rest and relative offsets are not supported.