Optimize the message window
scrollToBottom() was called on the table widget once for every message, which
triggered a new text layout every time, with O(n^2) complexity. Call it only
after receiving a batch of messages for hopefully going down to O(n)
complexity.

Also use the opportunity to switch to a more lightweight QListWidget and adjust
the themes accordingly.

See #584.