Speed up connecting to a game with many events
The code was still calling scrollToBottom() for every event because they are
sent one by one, resulting in O(n^2) complexity. Instead of calling it
directly, schedule a single call for when the server indicates that it's done
processing.

With this change, the layout of the message widget becomes invisible in
profiles.

See #584.