Also scroll for spontaneous messages
When the server sends a message outside of any request (e.g. when a unit is
killed by someone else in real time), there is no way to know when it will be
done sending, thus scroll immediately. This shouldn't be a large performance
hit since the server is unlikely to send hundreds of messages this way.