Don't crash when loading the list of scenarios
The client was crashing when loading a scenario, exiting the game, and trying
to go back to the scenario page. The reason was that the list of scenarios is
reloaded in this case, and the first time the loading code does is to clear the
QListWidget holding the list of scenarios. This propagated to the "selection
changed" callback that was trying to access the (nonexistant) first selected
item, resulting in a crash.

The fix is to check that there's something selected before trying to use it.

See #683.