Properly handle errors when reading spec files
I/O (or decompression) errors that occurred while reading spec files resulted
in an infinite loop. This was the case e.g. when trying to read an empty file
with ".gz" extension -- for instance, a failed save.

Make sure that any error is properly handled and terminates reading the file.
This makes the code a bit awful because the QIODevice error reporting API is
just terrible.