Allow falling back to a different AI type by default
Allow players to load scenarios (and saves) even if the chosen AI type is not
supported. In other words: "be liberal in what we accept."

In #746, trying to load a scenario using the "tex" AI failed because we don't
build it for fc21. If the user really wants another AI, s.he can switch back in
pregame. Scenarios that really want to lock down their AIs retain this
capability.