Fix integer overflow leading to division by zero
Taking the square of danger creates too large numbers in some cases. The
previous workaround - casting to long - is not guaranteed to work by the
standard and indeed it doesn't work on Windows.

This reverts commit c67fc6a2058028582ecdaca48d847bfdb4db1112. The succint
commit message there says that making danger quadratic is in line with defense.
But danger is already quadratic in assess_danger, so I guess that commit was a
mistake. In the worst case this restores the 2.6 formula, and AI didn't behave
that bad there.