Solve layout issues in the city screen right panel
heightForWidth() was reporting inconsistent results. Turns out it's not needed:
drop it.

Closes #665.
The symptoms described there (overlapping QLabels) sill happen when the window
is resized to a size too small for the contents, but this will be solved later
when we get rid of absolute positioning.  (This time, I hope, for good.)