Add tileset messages to the tileset debugger
Any message with a level < LOG_FATAL is now saved so the user can have a look
later on. This is in particular useful for terrain sprites, since they generate
LOG_ERROR messages when missing.

There are still many conditions under which a crash is the only solution.

Related to #632.
Related to #806.