Don't repeatedly show the tileset warning dialog when zooming
Zooming in or out reloads the tileset completely, and thus the warning would be
repeated when zooming in or out with a broken tileset. This would be too
obtrusive for the user since he was already warned once and zooming to the
desired level often requires several steps.

So never show a tileset error warning when zooming in or out. This is only
incorrect if someone is editing the tileset and zooming at the same time, which
should be rare compared to simple zoom actions, even for tileset authors.