Add a new (and basic) path finding implementation
The new implementation is based on the functions used by the server to tell
whether a move is possible. It finds correct paths for units without fuel, hp
loss or other fancy features. It also doesn't consider embarking or
disembarking.

Compared to the old implementation, this should correctly support
EFT_MOVE_BONUS, for instance in Royale units will avoid ending their turn on
mountains (the old behavior was to always end turns on mountains because of the
defense bonus).