Add fuel support to the new path finding
Fuel requires the use or several vertices per tile (bounded by 2 * fuel, but
the code doesn't need to know) to account for two possibilities for each fuel
level:

- Getting to the tile as quickly as possible (smallest number of turns)
- Getting to the tile with the largest number of moves left (to cross a long
  ocean, or perform an attack)

This requires the generation of a new kind of vertices: "rest" vertices where
the unit waits for one turn. They are generated for every tile, but only saved
when they result in a refuel. This is achieved by saying that certain pairs of
vertices are "not comparable", in which case they cannot be merged even if one
of the costs compares smaller.