Add support for HP loss and recovery in path finding
Helicopters will now seek cities to repair and stay there as long as needed
(based on currently available information: the time to repair will be
overestimated for allied cities with airports -- user responsibility). Similar
to fuel, sending units to their death is forbidden. Since these problems are
nearly identical, most of the code is shared.