Implement patrol support
For the server, patrols are normal goto path that repeat themselves and have
the "vigilant" flag set to break the orders when the unit uncovers something
interesting. There is no support for fuel or anything like that.

Finding a patrol path is identical to finding a path with a waypoint at the
destination and the end of the path at the current position of the unit.
Implement them exaclty in this way.

Because adding or removing waypoints currently resets the whole path finding
state, finding long patrol paths is quite slow -- but it works.