Add support for path final actions
The path finding algorithm will consider adding new nodes with ACTION_MOVE when
possible, letting the user choose the action to perform upon arrival. It uses
the exact same logic as the server (which may be buggy in some cases because it
wasn't written with the client in mind, even if for some reason the code is in
`common`).

This patch is very slow when there is no solution, because Dijkstra needs to
check every vertex before it can conclude that there is no solution.