Add support for empty paths
Previously, tiles returning empty paths were always seen as impossible
destinations. This is not the case for the tile the unit is on, which is of
course a possible destination even if the path is by construction empty.

Have find_path() return an std::optional<path> so it can signal when no path is
found.