Don't consider every ORDER_ACTION_MOVE a dangerous move
Vertices with ORDER_ACTION_MOVE are annoying because their move cost is unknown
and, as a consequence, one can't determine whether they will be safe fuel-wise.
The previous implementation considered them always unsafe, which resulted in
obvious mispredictions for units without fuel. With this patch, I switch to a
more optimistic heuristic: the vertex is (possibly) considered safe if it
involves an allied city or unit. Otherwise, it is still unsafe.