Add an automatic tool to run clang-format
By commenting on a PR with a message that contains "Auto clang-format",
maintainers and PR authors can trigger an automated run of clang-format.  This
can be used to fix PRs when their author isn't willing to do so or is simply
not responding.

The action posts a message reminding to avoid this for complex PRs: this action
is mainly intended at PRs with a single commit where the reformatting can be
squashed easily.  Anything more complex should be handled by a human.

Developed and tested in lmoureaux#1.
See #923, #947 for PRs that would [have] benifited from this action.
See #924 for an example where I think use of this action not indicated.