Implement a minimap panel at the bottom right
The panel contains the minimap and the Turn Done button. We lost the ability to
make it blink in the last seconds of the turn, which will be restored in a
further commit. The panel has its own .ui file so we can expand it easily later
on.

Resizing the minimap is no longer possible (but one can still hide it). Its
size depends on the map size and it is capped to some guessed size.

See #1002.