Have Autorevision create a PR instead of trying to push
Direct pushes to master are forbidden by branch protection rules. Have the
automation create a PR instead, so we can approve it and enable automerge. It's
a bit less convenient than a push but good enough for now. It also lets us
check that the action is working is intended. If this works well, we could
consider enabling automerge from within the action (or pushing directly as it
looks like it's supported via an access token [1]).

Closes #955.

[1] https://github.com/stefanzweifel/git-auto-commit-action/issues/87