Add a class to create colorized sprite variants
We need a method to make sprites depend on the player color (see #429). After
some research, it seems that a standard technique involves shifting pixels in
the image that have a specific hue. This is a quite simple method with very few
free parameters; one could imagine changing the saturation and lightness as
well to create "darker" or "more gray" nations.

Create a class that encapsulates all of this so it's easy to change later. It
generates colorized sprites on the fly (might create flicker as this is an
expensive operation...) and caches them for later use. The sprites are only
meant to be freed at tileset unload, so there is no way to clear the cache.

See #429.