Enable color replacements for unit sprites
Allow tilesets to specify a new configuration key, "replaced_hue", that
specifies the HSL hue that will be replaced by the hue of the player color in
unit sprites. A single hue is used for the whole tileset to facilitate the
implementation. The feature is disabled by default for backward compatibility
and none of the shipped tilesets uses it.

Closes #429.