Take into account that the minimap can be hidden
There is a menu item to hide the minimap. It was only toggling the minimap
without changing the layout of the panel at the bottom right, resulting in an
almost-empty widget still hiding as much of the map as before. Now, hiding the
minimap properly reduces the size of the bottom right widget, at the expense of
also hiding the zoom buttons located there.

The manual layout (i.e. without using QLayout) causes some trouble when widgets
are hidden or restored. Eventually I chose to post a LayoutRequest event by
hand to the widget doing the layout. If one day we write a proper QLayout for
the overlay widgets, we can remove this.

Closes #1096.