Update shortcuts in the menu when they change
The menu wasn't updated when the user changed a shortcut from the UI. Implement
it as well as we can without rebuilding the whole menu from scratch: look for
an action with the old shortcut, and change it to the new shortcut. It doesn't
work when the previous shortcut involves the mouse, as we can't guarantee that
we're changing the correct entry in that case.

See #1090.