Rewrite the shortcuts dialog
The way configurable shortcuts were stored had a number of bugs related to the
use of numerical key values. For instance, trying to set a shortcut to "+"
would result in "Shift++" (for keyboard "+") or "Num++" (for numpad "+"). This
shortcut was then only working when using Shift+numpad +, which is not what the
user entered.

Unfortunately, getting the "currently typed QKeySequence" is impossible in Qt.
The canonical (and only viable) way of letting the user enter a QKeySequence
without using or duplicating private Qt APIs is with QKeySequenceEdit.
Unfortunately QKeySequenceEdit is for key sequence only.

This commit rewrites the shortcut classes to:
- Store shortcuts as proper QKeySequences
- Use QKeySequenceEdit as the input widget, with some changes to allow for
  mouse input as well
- Fix various cases where duplicate shortcuts were misdetected

Not all functionality is available yet. In particular, shortcuts that are not
available through the menu cannot be triggered using the keyboard, and not all
shortcuts support being used with the mouse (this was already the case).

The file format of saved shortcuts has changed. No migration path is provided.

Closes #1090.