Manage menu actions shortcuts from fc_shortcuts
Walking through the menu to update action shortcuts, as was done previously, is
very brittle and doesn't work well with mouse shortcuts (since QAction doesn't
support them). This commit moves most of the shortcut handling logic to
fc_shortcuts: once an action is registered with it, its shortcut will be
updated automatically.

We still need to walk the menu to find shortcuts between hard-coded shortcuts
and user input, but that's much less likely to break since these shortcuts are
guaranteed to be keyboard only.