Fix keyboard shortcuts not in the menu
Not all shortcuts are mapped to actions in the menu. They stopped working when
we transitionned to QKeySequence. Restore this functionality by creating
QShortcut objects as needed and routing their signals to the mapctrl.cpp
function that was handling them.

This patch destroys mouse handling, but it will be restored in the next patch,
and in a better way.