Fix loading saved shortcuts
QSettings expects arrays to start at 0 and won't report the size otherwise. We
were starting it at 1 because of the unused SC_NONE as the first entry in the
shortcut_id enum. Remove it and shift all enumerators by one so the first is at
0.

This breaks compatibility with earlier versions, but we're using V2 now anyway.