Remove Q_UNREACHABLE()
It was used wrongly in many places and without proper documentation
as to why it is used. Remove it to make the warning showup
on coverity so that we can try to solve it the right way by refactoring the code.
and maybe put it back with right documentation as to why it is used.