Remove default_tileset_name
The last used tileset was saved in the options file, but that value wasn't
shown anywhere in the interface. It was then used *instead* of the tileset
configured in the settings when the topology matched. This was a source of
confusion. Get rid of this behavior by removing the hidden option.