Introduce map_updates_handler
This class will be used to implement the functionality of
refresh_tile|unit|city_mapcanvas and the like. It uses listener<> to provide a
bridge between the old C-style API and the newer object-oriented code.

So far, the class can accumulate a list of updates and restitute them.
Eventually it should be able to notify when a new frame is needed.