Make map_updates_handler notify when it has an update
Emit a signal whenever an update is ready, and connect it to
unqueue_mapview_updates. Use a queued connection so a few updates can
accumulate before they are processed in the next step of the event loop. This
allows us to remove the "idle callback" completely (a QueuedConnection is
basically the same thing, only in a more elegant way).