Remove center_tile_mapcanvas()
Calling into the map view widget directly means that we need to include
page_game.h and mapview.h in many places, but this was already the case with
mapview_common.h after all.

A small API change was needed to implement the behavior of can_slide.
Eventually some of it could be moved to the renderer, but the initial centering
will always need to bypass the animation.

Opening a city doesn't look as pretty as before because while
center_tile_mapcanvas() was synchronous, its replacement is inherently async.
We'll probably want to throw a QSequentialAnimationGroup at this problem.