page_load: Reserve some space for the info column
The column with game information in the page_load dialog (to the right of the
window) was still jumping around depending on the size of the map. There were
also concerns that the Longturn logo shown at the bottom right with the
Nightstalker theme was hidden (which still happens when the window is small,
but fixing that is a bigger change).

The column size is hacked in by having the right column expand to 25% of the
window width.

See #1166.