Use std::vector for the list of focused units
I would like to have some decent container to use in #805 (if several units are
selected, I want the right click menu to apply to all of them, which will
require a container).

Change focus_units (=get_units_in_focus()) to be a std::vector<unit *>. This
has implications in a lot of places, which are also modified to use
std::vector.