Prevent an assertion when observing
When, after starting a game, the user does /observe to his own player,
client.conn.playing is null for a brief period of time. While it's null, one
can't evaluate whether actions are valid or not because we don't know who we
are playing. This was triggering an assertion in update_actions(); just skip it
when we're not playing anyone (same as when we don't have a focused unit).

Closes #457.