Expand the city dialog text for military happiness
Improve the description of martial law by listing the number of units applying
it and the total number of citizens made happier as a result.
Greatly expand the description of military unhappiness by explaining the three
effects that come into play and providing numbers similiar to the above.

Should help with #269.