Fix horizontal edges not showing in iso-hex grids
In the initial implementation of gui_rect_iterator, it was assumed that each
iteration point corresponds to a single type of object in the geometry. This
was without counting on the laziness of Freeciv's hex implementation, which
simply collapsed one of the vertical edges into a corner. In hex mode, some
iteration points need to be corners and edges at the same time.

Closes #1253.