Simplify resizable_widget
Instead of letting the user of the class allow resizing of each corner
separately, only use the edges to specify which resizes are allowed. Usable
corners are deduced from the list of edges. This makes for a more concise
implementation and lets us use Qt::Edges instead of a custom enum.

Closes #1145.