Remove comments headers with too many stars
As a rule, we keep our function header comments simple:

/**
 * Some comment
 */

The two stars are for Doxygen to recognize a docstring.

There were still parts of the code with older, longer headers with many stars:

/***************************

or

/**********************/ /**

The first variant confuses Doxygen, and the second is just ugly (tm). I used
sed and a few regexes to edit all comments using the old variants. In some
cases, the comments were describing the whole file; in such cases I added \file
for Doxygen to pick them up.