Don't use the capital for the tooltip of the city top bar button
The tooltip of the "Cities" button in the top bar was showing the number of
cities by querying the help text for the player's capital. Now that we have
better help texts, this string is more specific about the capital and breaks in
the context of the top bar.

Change back the tooltip to what it was before, i.e. a simple city counter.