Don't default to huge fonts of HiDpi systems
We have enabled Qt's transparent handling of HiDpi, and thus the manual scaling
that was done in configure_fonts() was useless. Worse, on HiDpi diplays, it
would choose larger fonts to make the UI bigger, which Qt also does internally.
This resulted in huge fonts on such systems.

Simply delete the scaling code. Use the opportunity to tidy up a bit.

See #1273.