Set a baseline palette for all themes except System
In order to achieve consistency in what is displayed regardless of the system
theme, the system color scheme needs to be overriden. Previously only 3 colors
were changed (and for a single theme), which was causing readbility issues when
the system color scheme didn't match the theme.

Always fill the system QPalette with dark colors for dark themes and light
colors for light themes. Use the Breeze Light and Breeze Dark palettes as
references.

There are still some issues when changing the theme on the fly, much like #762.
It looks fine after restarting the game.