Add effect Wonder_Visible
This effect controls whether information about small wonders is visible to
players (and hence sent to the client). It lets one control it for each wonder
type separately, or possibly make it depend on the government (it is sadly not
possible to make it depend on the diplomatic state).

See #318.