Disconnect from server when going back to PAGE_MAIN
When cancelling a network connection then going back to PAGE_NETWORK, the old
connection was still around and any attempt to connect would fail with
"Connection in progress". Cancel the connection when switching back to
PAGE_MAIN so we can create a new one when we go back to PAGE_NETWORK.

Might be related to #785 or #1008.