Restore goto into the unknown
Add a flag to the path finder class to control whether it should include
unknown tiles in the search. Bind it to the corresponding gui option when
creating the finders.

This restores the ability to produce paths going through unknown parts of the
map, which is useful in some circumstances.

Closes #890.