Rewrite the client CMakeLists.txt
The CMakeLists.txt for the client/ subdir had been written before the client
code was consolidated into client/ (we still had client/gui-qt and a few
others). It also had the requirement of a separate libclient, which is no
longer needed.

Rewrite the file to create a single target (apart from source generation) with
all sources added when needed. Use the opportunity to group include directories
and linked libraries together.