Fix slanted lines in the minimap view rectangle
Tiny differences between the results of floating-point calculations that would
yield the exact same results in infinite precision culminated into off-by-one
errors in the calculation of the viewport coordinates on the minimap. Round the
floating-point results when converting them back to integers. This change might
lose some formal accuracy but the difference is invisible in practice.

Closes #1128.