Upload clang-tidy results to Codacy
Codacy has built-in support for clang-tidy, which is leveraged to produce
reports. This is achieved by running a two-steps Github Action on pull
requests. The first step, that runs without particular permissions, compiles
and runs clang-tidy. The second has more permissions and can use a secret API
key to upload the results to Codacy.