Add support for generic step constraints in path_finder
A step constraint is anything that limits the scope of the search by discarding
steps that would normally be considered. This can (and will) be used to
implement the "forbid goto into the unknown" at the client. Another use case
would be to limit the search depth (useful on the server).