Add a ruleset option for the initial diplomatic state
When two players meet for the first time, they're usually at war. Allow the
ruleset to override this with a state of their choosing. Use War for all
ruleset except for experimental, which will use Cease-fire, and Royale which
uses Armistice like in the Longturn Traditional rules.

This ports a76a6ce922eab1a22efb046078a05d0379dd068f from the Longturn 2.6 repo,
but changes it to be a ruleset setting.

See #1154.