Fix crash when plagues are enabled
The game was crashing when plagues were enabled (as in civ2civ3, used in the
tutorial). This was because the city screen was constructed with illness OFF
and it was later changed to ON.  Since the value of game.info.illness_on can
change at any time and I don't want to track it, simply set the visibility of
the plague labels when updating the city screen and plague is disabled.

Closes #1473.