Fix a bug where the Messages widget could be moved
It was possible to grab a corner of the Messages widget and move it in a
direction in which moves supposedly were disallowed (for the top left corner,
up or down; for the bottom right corner, left or right), effectively moving the
widget out of the corner. Fix the logic used to determine which edges can be
moved in a drag-and-drop.