Don't try to redefine (bind)textdomain
Defining `textdomain' as an empty macro conflicts the inclusion of `libintl.h'
when it defines `textdomain' as a function. This becomes problematic with
transient includes of `libintl.h' from standard library headers not under our
control: for instance, <locale> may implement std::messages using inline
gettext calls.

This patch fixes a compilation error with GCC 11 when NLS is disabled.