Sort rows in the city dialog detailed output
Sort and categorize the table showing the detailed output of cities. The four
main categories are Food, Production, and Trade, in the order they are shown on
the map. Further details are shown under each category:

- Under Food: everything related to growth
- Under Production: only waste (the difference between waste and corruption is
  hard to learn, maybe this will help...)
- Under Trade: corruption and the three outputs taxed from it, in the order
  they are shown in the national budget.

This is only a first pass: in the future we'll want to add icons, make this
more colorful, and get rid of the monospace font.

Closes #1491.