Fix saving map images
Since the new renderer class is asynchronous, we need to wait until the first
repaint notification arrives before saving the image. This required a rather
big restructuring of the function, but it should do exactly the same as it used
to.

We still need to fiddle with map_canvas_resized() because the mapview code only
supports a single canvas shared between all renderers (the "renderer"
abstraction is, truly, a lie).

Closes #1486.