Always update enum and bitwise options when the server sends them
The options dialog used not to be refreshed because of a logic error in
handle_server_setting_bitwise(), which caused the old values to remain
displayed and other fancy behavior. Make sure to always call
option_gui_update(), even if it would maybe be redundant -- better waste a few
CPU cycles than display wrong information.

Closes #1122.