Restore old tech tree colors coming from the tileset
This mostly reverts e9a0a32ce5f9ccbf5ad22b8a370d9339aad4cdfb, because attempts
at using it with Classic ended up badly. In addition to undoing these changes,
this commit adds the ability to override colors in the theme configuration.
Nightstalker colors will be restored in an upcoming commit.

Eventually we should migrate all out-of-map colors to the themes. For now, rely
on the nice-looking (with Classic) colors coming from the tilesets.

Part of #1517.