Prevent the autocenter option from autochanging
The "auto center on unit" option was being changed without user interaction
because of badly engineered code that was temporarily changing its value and
failing to restore it. Make the code safer by introducing a new, purely
internal, flag to pause the auto centering so we don't need to touch the
user-visible setting.

Closes #1316.