Take team_pooled_research into account when looping over research
The number of valid research objects is the number of teams when
team_pooled_research is enabled, and the number of players when it is disabled.
The code was always assuming the first case, which created bugs in team games
without pooled research.

Fixes 22c0c4b47ed3ed239358933af15818c7b4873c49.