Always send wonder information to the player's team
With the introduction of EFT_WONDER_VISIBLE, it became possible not to send
(small) wonder information to the team owning them. This is obviously not the
right thing to do, since the players can see the wonders in the cities. So
always send the info.

Closes #1656.

(cherry picked from commit c6200fc378c9a17bd27f7683d29898d95d68da42)